pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "text-to-sql"
version = "0.1.0"
description = "Text-to-SQL POC with Claude API and SQLite MCP"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require ANTHROPIC_API_KEY and external services",
]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
